+++ src/post_tree.rs
@@ -3,12 +3,12 @@ use crate::internal::*;
 pub fn post_tree(
     root: i32,
     mut k: i32,
-    child: &mut [i32], // input of size nn, undefined on output
-    sibling: &[i32],   // input of size nn, not modified
-    order: &mut [i32], // output of size nn
+    child: &mut [isize], // input of size nn, undefined on output
+    sibling: &[isize],   // input of size nn, not modified
+    order: &mut [isize], // output of size nn
     nn: i32,
 ) -> i32 {
-    let mut stack: Vec<i32> = vec![0; nn as usize];
+    let mut stack: Vec<isize> = vec![0; nn as usize];
 
     /*if false {
         // Recursive version (stack[] is not used):
@@ -28,14 +28,14 @@ pub fn post_tree(
 
     // Push root on the stack.
     let mut head = 0;
-    stack[0] = root;
+    stack[0] = root as isize;
 
     while head >= 0 {
         // Get head of stack.
         debug_assert!(head < nn);
         let i = stack[head as usize];
         debug1_print!("head of stack {} \n", i);
-        debug_assert!(i >= 0 && i < nn);
+        debug_assert!(i >= 0 && i < nn as isize);
 
         if child[i as usize] != EMPTY {
             // The children of i are not yet ordered
@@ -46,7 +46,7 @@ pub fn post_tree(
             while f != EMPTY {
                 head += 1;
                 debug_assert!(head < nn);
-                debug_assert!(f >= 0 && f < nn);
+                debug_assert!(f >= 0 && f < nn as isize);
                 f = sibling[f as usize];
             }
             let mut h = head;
@@ -57,7 +57,7 @@ pub fn post_tree(
                 stack[h as usize] = f;
                 h -= 1;
                 debug1_println!("push {} on stack", f);
-                debug_assert!(f >= 0 && f < nn);
+                debug_assert!(f >= 0 && f < nn as isize);
                 f = sibling[f as usize];
             }
             debug_assert!(stack[h as usize] == i);
@@ -69,7 +69,7 @@ pub fn post_tree(
             // remove i from the stack and order it. Front i is kth front.
             head -= 1;
             debug1_print!("pop {} order {}\n", i, k);
-            order[i as usize] = k;
+            order[i as usize] = k as isize;
             k += 1;
             debug_assert!(k <= nn);
         }
